notaris & cluster master after rework

+++ app/Domains/Master/Http/Backend/NotarisController.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace App\Domains\Notaris\Http\Backend\Controllers;
+namespace App\Domains\Master\Http\Backend\Controllers;
 
 use App\Controllers\Requests\BaseRequest;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Backend\BaseBackendController;
-use App\Domains\Notaris\Services\NotarisService;
-use App\Domains\Notaris\Http\Requests\NotarisRequest;
-use App\Domains\Notaris\Models\Notaris;
+use App\Domains\Master\Services\NotarisService;
+use App\Domains\Master\Http\Requests\NotarisRequest;
+use App\Domains\Master\Models\Notaris;
 
 class NotarisController extends BaseBackendController
 {
@@ -27,7 +27,6 @@ class NotarisController extends BaseBackendController
      */
     public function store(NotarisRequest $request)
     {
-
         $this->service->store($request->validated());
 
         return redirect()->route($this->route_view_index)->withFlashSuccess(__('The Data was successfully created.'));
@@ -40,6 +39,9 @@ class NotarisController extends BaseBackendController
 
     public function update(NotarisRequest $request, Notaris $notaris)
     {
+        if($request->partner_id == $notaris->id) {
+            return back()->withErrors('Notaris dan partner tidak boleh sama');
+        }
         $this->service->update($notaris, $request->validated());
 
         return redirect()->route($this->route_view_index)->withFlashSuccess(__('The Data was successfully updated.'));
@@ -47,6 +49,10 @@ class NotarisController extends BaseBackendController
 
     public function destroy(BaseRequest $request,Notaris $notaris)
     {
+        $partnerId = Notaris::where('partner_id', '=', $notaris->id)->get();
+        if(!empty($partnerId)) {
+            return back()->withErrors('Notaris digunakan sebagai partner');
+        }
         $this->service->destroy($notaris);
 
         return redirect()->route($this->route_view_index)->withFlashSuccess(__('The Data was successfully deleted.'));
